+++ bin/api/start-api-dev.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+python3 -debug -verbose -verbose -verbose api