+++ bin/start-api-dev.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-python ./OnlySnarf/api/index.py -debug -verbose -verbose -verbose users