+++ bin/update-and-start.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 /usr/bin/pip install -U onlysnarf
-/usr/local/bin/snarf -debug -verbose -verbose -verbose api
+/usr/local/bin/snarf api