+++ scripts/download-from-dropbox.py
@@ -1,0 +1,5 @@
+# script that downloads from dropbox
+
+# download from dropbox:
+# /OnlySnarf/post 		--> 	~/.onlysnarf/uploads/post
+# /OnlySnarf/message 		--> 	~/.onlysnarf/uploads/message