+++ scripts/scan.py
@@ -1,0 +1,7 @@
+# i need to configure a script for sending posts & messages via json snippets left in folders
+
+# - read each directories for json file
+# if json file,
+# - check json file for proper metadata
+# else, use default data
+# - post / message via OnlySnarf